Conditional S3 GET requests for files < 5GB.
Change the artefact download so that it makes a conditional GET request using the 'If-None-Match' request header if the local file exists and is no greater than 5GB. Files greater than 5GB are not checked this way and retain the old behavior, since such files have an ETag that does not solely consist of the MD5 digest due to multipart-upload. Note that the conditional GET will also fail (and thus re-download the file every time) if it is < 5GB but has been uploaded through the multipart upload API anyway.

Addresses https://github.com/khan/issues/127 and makes using '--force' in hegemony unnecessary in most cases.