Bring back the routes subcommand
This turned out to be needed after all. Since we only use it on klabautermann, this PR also changes it to emit JSON suitable for input to an ed-e template.

The advantage of this approach is that we avoid the ed-e dependency and the HAProxy template in khan, permitting klabauternamm to perform the expansion itself.

One drawback is that the routes subcommand now produces a potentially large chunk of unformatted JSON. This could be improved upon by having it (by default) produce nicely formatted routes in plain text, but the use case for this seems marginal and it can easily be added later.